backend/requirements.txt
# ============================================
# IKVCS 后端核心依赖
# 适用于：快速开发 + Postman API 测试
# ============================================

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1
hiredis==2.2.3

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP 客户端（调用 LLM API）
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
python-dateutil==2.8.2

# 定时任务（Redis 到 MySQL 数据同步）
apscheduler==3.10.4
